discord_crypto_bot_1/requirements.txt
python-dotenv
schedule
requests
psycopg2
py-cord
loguru